+++ courses-app/src/App_sena_versija.js
@@ -4,7 +4,7 @@ import { NavLink } from "react-router-dom";
 import { Route, Switch, BrowserRouter } from "react-router-dom";
 import Courses from "./Courses";
 import Home from "./Home";
-import CourseCards from "./Models/CourseCard";
+import CourseCards from "./Components/CourseCard";
 import InfoAboutCourse from "./InfoAboutCourse";
 import Filter from "./Filter";
 import Header from "./Header";
@@ -14,63 +14,63 @@ class App extends React.Component {
   render() {
     const Search = () => {
       return (
-      <div>
-       
+        <div>
+
           <iframe width="420" height="315" src="https://www.youtube.com/embed/tgbNymZ7vqY?autoplay=1" />
-        
-        <div class="example">
-          <input type="text" placeholder="Search.." name="search" />
-          <button type="submit">Search</button>
-        </div>
-    
-      </div>);
-    }
-      return (
-          <div>
-              <Header />
-             Header...
-             <BrowserRouter>
-    {/* Search start */}
-    <div><Search /></div>
-    {/* Search end */}
-    <div>
 
-      <div className="NavContainer">
-        <button className="NavButton">
-          <NavLink to="/home">Home</NavLink>
-        </button>
-
-        <button className="NavButton">
-          <NavLink to={{ pathname: "/courses" }}>Link to courses</NavLink>
-        </button>
-
-        <button className="NavButton">
-          <NavLink to={{ pathname: "/courseCards" }}>Link to courses</NavLink>
-        </button>
-
-        <button className="NavButton">
-          <NavLink to={{ pathname: "/infoAboutCourse" }}>Info about course</NavLink>
-        </button>
-        <button className="NavButton">               
-                <NavLink to={{ pathname: "/filter" }}>Filter component</NavLink>
-            </button>
-
-      </div>
-      <Switch>
-        <Route exact path="/courses" component={Courses} />
-        <Route exact path="/home" component={Home} />
-        <Route exact path="/courseCards" component={CourseCards.CourseCards} />
-        <Route exact path="/infoAboutCourse" component={InfoAboutCourse} />
-        <Route exact path="/filter" component={Filter} /> 
-      </Switch>
-    </div>
-  </BrowserRouter>
-
-              <Footer />
+          <div class="example">
+            <input type="text" placeholder="Search.." name="search" />
+            <button type="submit">Search</button>
           </div>
-      )
+
+        </div>);
+    }
+    return (
+      <div>
+        <Header />
+        Header...
+             <BrowserRouter>
+          {/* Search start */}
+          <div><Search /></div>
+          {/* Search end */}
+          <div>
+
+            <div className="NavContainer">
+              <button className="NavButton">
+                <NavLink to="/home">Home</NavLink>
+              </button>
+
+              <button className="NavButton">
+                <NavLink to={{ pathname: "/courses" }}>Link to courses</NavLink>
+              </button>
+
+              <button className="NavButton">
+                <NavLink to={{ pathname: "/courseCards" }}>Link to courses</NavLink>
+              </button>
+
+              <button className="NavButton">
+                <NavLink to={{ pathname: "/infoAboutCourse" }}>Info about course</NavLink>
+              </button>
+              <button className="NavButton">
+                <NavLink to={{ pathname: "/filter" }}>Filter component</NavLink>
+              </button>
+
+            </div>
+            <Switch>
+              <Route exact path="/courses" component={Courses} />
+              <Route exact path="/home" component={Home} />
+              <Route exact path="/courseCards" component={CourseCards.CourseCards} />
+              <Route exact path="/infoAboutCourse" component={InfoAboutCourse} />
+              <Route exact path="/filter" component={Filter} />
+            </Switch>
+          </div>
+        </BrowserRouter>
+
+        <Footer />
+      </div>
+    )
   }
-} 
+}
 
 export default App;
