+++ courses-app/src/Components/Lector.js
